refactor: :recycle: Refactor mobile chat layout
