+++ src/components/chat/styles.module.css
@@ -124,6 +124,14 @@
   padding: 10px;
 }
 
+.tab {
+  padding: 12px 0px;
+  min-width: auto;
+  text-transform: initial;
+  margin: 0 8px;
+  font-size: 16px;
+}
+
 @media (min-width: 600px) {
   .hideDesktop {
     display: none;
