+++ src/pages/chat.tsx
@@ -1,12 +1,9 @@
 import { ChatOverview } from '@/components/chat/chatOverview'
 import { DesktopChat } from '@/components/chat/desktopChat'
 import { MobileChat } from '@/components/chat/mobileChat'
-import { AddFolderModal } from '@/components/chat/modals/AddFolderModal'
-import ViewCreateSafe from '@/components/chat/modals/CreateSafe'
 import ViewSettingsModal from '@/components/chat/modals/ViewSettingsModal'
+import { SafeList } from '@/components/chat/SafeList'
 import ConnectionCenter from '@/components/common/ConnectWallet/ConnectionCenter'
-import { FolderList } from '@/components/folder-list'
-import FolderGroup from '@/components/folder-list/folderGroups'
 import { AppRoutes } from '@/config/routes'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import useTxHistory from '@/hooks/useTxHistory'
@@ -14,31 +11,23 @@ import useTxQueue from '@/hooks/useTxQueue'
 import useWallet from '@/hooks/wallets/useWallet'
 import ellipsisAddress from '@/utils/ellipsisAddress'
 import { ArrowBackIos } from '@mui/icons-material'
-import AddIcon from '@mui/icons-material/Add'
 import SettingsIcon from '@mui/icons-material/Settings'
 import ViewSidebarIcon from '@mui/icons-material/ViewSidebar'
 import {
   Avatar,
-  Box,
-  Button,
-  Container,
+  Box, Container,
   Divider,
   Drawer,
   Hidden,
-  IconButton,
-  Tab,
-  Tabs,
-  Toolbar,
-  Typography
+  IconButton, Toolbar,
+  Typography,
+  useMediaQuery
 } from '@mui/material'
 import { styled } from '@mui/material/styles'
-import { getSession, signOut } from 'next-auth/react'
+import { getSession } from 'next-auth/react'
 import Head from 'next/head'
 import Link from 'next/link'
 import React, { useCallback, useEffect, useRef, useState } from 'react'
-
-import { useRouter } from 'next/router'
-import css from './styles.module.css'
 
 const drawerWidth = 360
 
@@ -59,40 +48,6 @@ const Main = styled('div', { shouldForwardProp: (prop) => prop !== 'open' })<{
     marginRight: 0,
   }),
 }))
-
-interface TabPanelProps {
-  children?: React.ReactNode
-  index: number
-  value: number
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  )
-}
-
-//wtf is this lol
-function a11yProps(index: number) {
-  return {
-    id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`,
-  }
-}
 
 //Get auth session, if not reroute
 export async function getServerSideProps(context: any) {
@@ -116,16 +71,12 @@ export async function getServerSideProps(context: any) {
 const Chat: React.FC<{
   user: any
 }> = ({ user }) => {
-  const router = useRouter()
+  const matches = useMediaQuery('(max-width: 600px)')
   //folders and folder control
   const [group, setGroup] = useState<any>()
-  const [folders, setFolders] = useState([])
-  const [popup, togglePopup] = useState<boolean>(false)
-  //modals and modal control
-  const [createSafe, setCreateSafe] = useState<boolean>(false)
   const [settings, toggleSettings] = useState<boolean>(false)
   const [open, setOpen] = useState(true)
-  const [value, setValue] = React.useState(0)
+
   //Chat
   const [message, setMessage] = useState('')
   const [messages, setMessages] = useState([''])
@@ -140,9 +91,6 @@ const Chat: React.FC<{
   const bottom = useRef<HTMLDivElement>(null)
   const owners = safe?.owners || ['']
   const ownerArray = owners.map((owner) => owner.value)
-  const resetGroup = () => {
-    setGroup('')
-  }
 
   const scrollToBottom = useCallback(() => {
     if (!bottom.current) return
@@ -153,34 +101,6 @@ const Chat: React.FC<{
     }
     bottomOfChat.scrollIntoView({ behavior: 'smooth' })
   }, [])
-
-  useEffect(() => {
-    if (user.address !== wallet?.address) {
-      //@ts-ignore
-      signOut({ redirect: '/auth' })
-    }
-    if (router.asPath.includes('chain')) {
-      setCreateSafe(true)
-    }
-  }, [])
-
-  useEffect(() => {
-    const activeFolders = async () => {
-      const items = JSON.parse(localStorage.getItem('folders')!)
-      if (items) {
-        setFolders(items)
-      }
-    }
-    activeFolders()
-    window.addEventListener('storage', activeFolders)
-    return () => {
-      window.removeEventListener('storage', activeFolders)
-    }
-  }, [])
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue)
-  }
 
   const toggleDrawer = (open: boolean) => (event: React.KeyboardEvent | React.MouseEvent) => {
     if (
@@ -279,9 +199,7 @@ const Chat: React.FC<{
 
   return (
     <>
-      {popup && <AddFolderModal open={popup} onClose={() => togglePopup(!popup)} />}
       {settings && <ViewSettingsModal open={settings} onClose={() => toggleSettings(!settings)} />}
-      {createSafe && <ViewCreateSafe open={createSafe} onClose={() => setCreateSafe(!createSafe)} />}
       <Head>
         <title>Decentra &mdash; Chat</title>
       </Head>
@@ -304,40 +222,7 @@ const Chat: React.FC<{
             variant="permanent"
             anchor="left"
           >
-            <Toolbar sx={{ display: 'flex', justifyContent: 'space-between' }}>
-              <Box>
-                <Typography sx={{ color: '#757575', fontSize: 12, fontWeight: 600 }}>VIEW AS:</Typography>
-                <Typography sx={{ fontWeight: 600 }}>{ellipsisAddress(`${wallet.address}`)}</Typography>
-              </Box>
-              <Box display="flex" alignItems="center" gap="10px">
-                <IconButton
-                  sx={{ border: '1px solid var(--color-border-light)', borderRadius: '6px', width: '32px', height: '32px' }}
-                  aria-label="add folder"
-                  onClick={() => togglePopup(!popup)}
-                >
-                  <AddIcon />
-                </IconButton>
-              </Box>
-            </Toolbar>
-            <Box sx={{ width: '100%', height: '100%' }}>
-              <Tabs sx={{ padding: '0 16px', borderBottom: '1px solid var(--color-border-light)' }} value={value} onChange={handleChange} aria-label="folder tabs">
-                <Tab className={css.tab} label="All" {...a11yProps(0)} />
-                {folders.map((folder, i) => {
-                  return <Tab className={css.tab} label={folder} key={`${folder}-${i}`} />
-                })}
-              </Tabs>
-              <TabPanel value={value} index={0}>
-                <FolderList resetGroup={resetGroup} key={wallet?.chainId} />
-              </TabPanel>
-              {folders.map((folder, i) => {
-                return (
-                  <TabPanel value={value} index={i + 1} key={`${folder}-${i}`}>
-                    <FolderGroup group={folder} currentSafe={safeAddress} />
-                  </TabPanel>
-                )
-              })}
-              <Button onClick={() => setCreateSafe(!createSafe)}>Add Safe</Button>
-            </Box>
+            <SafeList user={user} />
           </Drawer>
         </Hidden>
         <Main open={open} sx={{ flexGrow: 1, bgcolor: 'var(--color-background-lightcolor)' }}>
@@ -356,11 +241,13 @@ const Chat: React.FC<{
                 }}
               >
                 <Box sx={{ display: 'flex', justifyContent: 'flex-start', alignItems: 'center', gap: '14px' }}>
-                  <Link href={{ pathname: AppRoutes.home, query: { safe: `${safeAddress}` } }}>
-                    <IconButton  className={css.hideDesktop} aria-label="back">
-                      <ArrowBackIos />
-                    </IconButton>
-                  </Link>
+                  {matches &&
+                    <Link href={{ pathname: AppRoutes.safeList, query: { safe: `${safeAddress}` } }}>
+                      <IconButton aria-label="back">
+                        <ArrowBackIos />
+                      </IconButton>
+                    </Link>
+                  }
                   <Avatar sx={{ height: 32, width: 32, borderRadius: '6px' }} alt="Decentra" />
                   <Typography sx={{ fontWeight: 600 }}>{safeAddress ? ellipsisAddress(`${safeAddress}`) : ''}</Typography>
                 </Box>
@@ -397,32 +284,36 @@ const Chat: React.FC<{
                   </Container>
                   :
                   <>
-                    <MobileChat
-                      message={message}
-                      setMessage={setMessage}
-                      messages={messages}
-                      setMessages={setMessages}
-                      bottom={bottom}
-                      chatData={chatData}
-                      group={group}
-                      owners={owners}
-                      currentUser={currentUser}
-                      setCurrentUser={setCurrentUser}
-                      setGroup={setGroup}
-                    />
-                    <DesktopChat
-                      setGroup={setGroup}
-                      currentUser={currentUser}
-                      setCurrentUser={setCurrentUser}
-                      message={message}
-                      setMessage={setMessage}
-                      messages={messages}
-                      setMessages={setMessages}
-                      group={group}
-                      bottom={bottom}
-                      chatData={chatData}
-                      safe={safeAddress}
-                    />
+                    {matches &&
+                      <MobileChat
+                        message={message}
+                        setMessage={setMessage}
+                        messages={messages}
+                        setMessages={setMessages}
+                        bottom={bottom}
+                        chatData={chatData}
+                        group={group}
+                        owners={owners}
+                        currentUser={currentUser}
+                        setCurrentUser={setCurrentUser}
+                        setGroup={setGroup}
+                      />
+                    }
+                    {!matches &&
+                      <DesktopChat
+                        setGroup={setGroup}
+                        currentUser={currentUser}
+                        setCurrentUser={setCurrentUser}
+                        message={message}
+                        setMessage={setMessage}
+                        messages={messages}
+                        setMessages={setMessages}
+                        group={group}
+                        bottom={bottom}
+                        chatData={chatData}
+                        safe={safeAddress}
+                      />
+                    }
                   </>
               }
             </Box>
